update,refactor,test: removed file-transporter service and moved the logic of uploading a file to S3 in S3ManagerService
