+++ src/core/s3-manager/@types/index.ts
@@ -1,1 +1,2 @@
 export * from './interface';
+export * from './types';
