+++ src/core/s3-manager/@types/types.ts
